(C++) utils : Rename sum_mult.cpp to sum.cpp
This file will contain functions for different types of sums.
The functions also is now using templates, so that they can accept any
integral type. It can also now accept negative maximum values, that is,
the sum of the value (exclusive) up to 0 (which is the same as the negation
of the positive sums).